requirements.txt
google-cloud-discoveryengine
google-api-core
google-cloud-storage